+++ resources/views/admin/operatingsystem/create.blade.php
@@ -3,5 +3,7 @@
 
         <x-create.singlerow :label="__('Name')" name="name" />
 
+        <x-create.singlerow :label="__('Support-Ende (EOL)')" name="eol_date" type="date" />
+
     </x-create.main>
 </x-admin-layout>