+++ resources/views/admin/operatingsystem/edit.blade.php
@@ -4,5 +4,7 @@
 
         <x-create.singlerow :label="__('Name')" name="name" :default="$operatingSystem->name" />
 
+        <x-create.singlerow :label="__('Support-Ende (EOL)')" name="eol_date" type="date" :default="$operatingSystem->eol_date?->format('Y-m-d')" />
+
     </x-create.main>
 </x-admin-layout>